bugfix: The device inactivity condition was not handled correctly.
The handling of device inactivity wasn’t implemented properly earlier. When the device is inactive, the request:

{{ $currentlyPlaying := newRequest "https://api.spotify.com/v1/me/player"
                    | withHeader "Authorization" (print "Bearer " $accessToken)
                    | getResponse
                }}

returns an empty response with a 204 status code.

To address this, I’ve added a new check for this specific condition. Previously, since this case wasn’t accounted for, the widget would display nothing whenever the device was inactive.